Rename and move to toposort package
