+++ fabric-api-base/src/main/java/net/fabricmc/fabric/impl/base/event/EventPhaseData.java
@@ -21,6 +21,8 @@ import java.util.Arrays;
 
 import net.minecraft.util.Identifier;
 
+import net.fabricmc.fabric.impl.base.toposort.SortableNode;
+
 /**
  * Data of an {@link ArrayBackedEvent} phase.
  */
